+++ lib/link.bash
@@ -54,11 +54,12 @@ function link_is_addon_linked {
 }
 
 
-# link_module_impl <source_path> <dest_path> <force: on|off>
+# link_module_impl <source_path> <dest_path> <force: on|off> <py-deps-manifest: on|off>
 function link_module_impl {
     local src; src=$(readlink -f "$1");
     local dest=$2;
     local force=$3;
+    local py_deps_manifest="${4:-off}";
 
     if [ "$force" == "on" ] && { [ -e "$dest" ] || [ -L "$dest" ]; }; then
         echov "Rewriting module $dest...";
@@ -81,59 +82,109 @@ function link_module_impl {
     fetch_requirements "$dest";
     fetch_pip_requirements "$dest";
     fetch_oca_requirements "$dest";
+
+    if [ -n "$py_deps_manifest" ] && [ -f "$dest/__manifest__.py" ]; then
+        local py_deps;
+        py_deps=$(exec_py "$ODOO_HELPER_LIB/pylib/py_utils.py" addon-py-deps --addon-path "$dest");
+        if [ -n "$py_deps" ]; then
+            odoo-helper pip install "$py_deps";
+        fi
+    fi
 }
 
-# link_module <force: on|off> <repo_path> [<module_name>]
+# signature:
+#     link_module [options] <repo_path> [<module_name>]
+# options:
+#     --force
+#     --module-name <module name>
+#     --fetch-manifest-py-deps
 function link_module {
-    local force=$1;
-    local REPO_PATH=$2;
-    local MODULE_NAME=$3
+    local force=off;
+    local fetch_manifest_py_deps=off;
+    local module_name;
 
-    if [ -z "$REPO_PATH" ]; then
-        echo -e "${REDC}Bad repo path for link: ${YELLOWC}${REPO_PATH}${NC}";
+    # Parse command line options and run commands
+    if [[ $# -lt 1 ]]; then
+        echo "No options supplied $#: $*";
+        return 1;
+    fi
+
+    # Process all args that starts with '-' (ie. options)
+    while [[ $1 == -* ]]
+    do
+        local key="$1";
+        case $key in
+            -f|--force)
+                force=on;
+            ;;
+            --fetch-manifest-py-deps)
+                fetch_manifest_py_deps=on;
+            ;;
+            --module-name)
+                module_name="$2";
+                shift;
+            ;;
+            *)
+                echo "Unknown option $key";
+                return 1;
+            ;;
+        esac
+        shift
+    done
+    local repo_path="$1"
+
+    if [ -z "$repo_path" ]; then
+        echo -e "${REDC}Bad repo path for link: ${YELLOWC}${repo_path}${NC}";
         return 2;
     fi
 
-    REPO_PATH=$(readlink -f "$2");
+    repo_path=$(readlink -f "$repo_path");
 
     local recursion_key="link_module";
-    if ! recursion_protection_easy_check "$recursion_key" "${REPO_PATH}__${MODULE_NAME:-all}"; then
-        echo -e "${YELLOWC}WARN${NC}: REPO__MODULE ${REPO_PATH}__${MODULE_NAME:-all} already had been processed. skipping...";
+    if ! recursion_protection_easy_check "$recursion_key" "${repo_path}__${module_name:-all}"; then
+        echo -e "${YELLOWC}WARN${NC}: REPO__MODULE ${repo_path}__${module_name:-all} already had been processed. skipping...";
         return 0
     fi
 
-    echov "Linking module $REPO_PATH [$MODULE_NAME] ...";
+    echov "Linking module $repo_path [$module_name] ...";
 
     # Guess repository type
-    if is_odoo_module "$REPO_PATH"; then
+    if is_odoo_module "$repo_path"; then
         # single module repo
         local basename_repo;
-        basename_repo=$(basename "$REPO_PATH");
-        link_module_impl "$REPO_PATH" "$ADDONS_DIR/${MODULE_NAME:-$basename_repo}" "$force";
+        basename_repo=$(basename "$repo_path");
+        link_module_impl "$repo_path" "$ADDONS_DIR/${module_name:-$basename_repo}" "$force" "$fetch_manifest_py_deps";
     else
         # multi module repo
-        if [ -z "$MODULE_NAME" ]; then
+        if [ -z "$module_name" ]; then
             # Check for requirements files in repository root dir
-            fetch_requirements "$REPO_PATH";
-            fetch_pip_requirements "$REPO_PATH";
-            fetch_oca_requirements "$REPO_PATH";
+            fetch_requirements "$repo_path";
+            fetch_pip_requirements "$repo_path";
+            fetch_oca_requirements "$repo_path";
 
             # No module name specified, then all modules in repository should be linked
-            for file in "$REPO_PATH"/*; do
+            for file in "$repo_path"/*; do
                 local base_filename;
                 base_filename=$(basename "$file");
                 if is_odoo_module "$file" && addons_is_installable "$file"; then
                     # link module
-                    link_module_impl "$file" "$ADDONS_DIR/$base_filename" "$force";
+                    link_module_impl "$file" "$ADDONS_DIR/$base_filename" "$force" "$fetch_manifest_py_deps";
                 elif [ -d "$file" ] && ! is_odoo_module "$file" && [ "$base_filename" != 'setup' ]; then
                     # if it is directory but not odoo module,
                     # and not 'setup' dir, then recursively look for addons there
-                    link_module "$force" "$file";
+                    local link_module_opts=( );
+                    if [ "$force" == on ]; then
+                        link_module_opts+=( --force )
+                    fi
+                    if [ "$fetch_manifest_py_deps" == on ]; then
+                        link_module_opts+=( --fetch-manifest-py-deps );
+                    fi
+                    link_module "${link_module_opts[@]}" "$file";
                 fi
             done
         else
             # Module name specified, then only single module should be linked
-            link_module_impl "$REPO_PATH/$MODULE_NAME" "$ADDONS_DIR/$MODULE_NAME" "$force";
+            link_module_impl "$REPO_PATH/$MODULE_NAME" "$ADDONS_DIR/$MODULE_NAME" "$force" "$fetch_manifest_py_deps";
         fi
     fi
 }
@@ -143,15 +194,14 @@ function link_command {
     local usage="
     Usage: 
 
-        $SCRIPT_NAME link [-f|--force] <repo_path> [<module_name>]
+        $SCRIPT_NAME link [options] <repo_path>
 
     Options:
-        -f|--force   - rewrite links if already exists
-        --ual        - update addons list after link
+        -f|--force                - rewrite links if already exists
+        --fetch-manifest-py-deps  - fetch python dependencies from addon's manifest
+        --module-name <name>      - name of module to link from repo
+        --ual                     - update addons list after link
     ";
-
-    local force=off;
-    local ual;
 
     # Parse command line options and run commands
     if [[ $# -lt 1 ]]; then
@@ -161,6 +211,7 @@ function link_command {
         return 0;
     fi
 
+    local link_module_opts=( )
     # Process all args that starts with '-' (ie. options)
     while [[ $1 == -* ]]
     do
@@ -171,7 +222,14 @@ function link_command {
                 return 0;
             ;;
             -f|--force)
-                force=on;
+                link_module_opts+=( --force );
+            ;;
+            --fetch-manifest-py-deps)
+                link_module_opts+=( --fetch-manifest-py-deps );
+            ;;
+            --module-name)
+                link_module_opts+=( --module-name "$2" );
+                shift;
             ;;
             --ual)
                 ual=1;
@@ -184,7 +242,7 @@ function link_command {
         shift
     done
 
-    link_module "$force" "$@";
+    link_module "${link_module_opts[@]}" "$@";
 
     if [ -n "$ual" ]; then
         addons_update_module_list;